feat(searchCommandMenu): refactor search functionality and add SearchCourseList component for improved course searching experience
