+++ components/shared/SearchCommandMenu.tsx
@@ -2,38 +2,23 @@
 import { Button } from '@/components/ui/button';
 import {
   CommandDialog,
-  CommandEmpty,
-  CommandGroup,
   CommandInput,
-  CommandItem,
   CommandList
 } from '@/components/ui/command';
 import { Skeleton } from '@/components/ui/skeleton';
-import { useGetCourses } from '@/hooks/course';
 import { useDebounce } from '@/hooks/useDebounce';
-import { cn, formUrlQuery, isMacOs, removeKeysFromQuery } from '@/lib/utils';
+import { formUrlQuery, isMacOs, removeKeysFromQuery } from '@/lib/utils';
 import { Search } from 'lucide-react';
-import Image from 'next/image';
 import { useRouter, useSearchParams } from 'next/navigation';
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, Suspense } from 'react';
+import SearchCourseList from '../sections/SearchCourseList';
 
 const SearchCommandMenu = () => {
   const router = useRouter();
   const searchParams = useSearchParams();
-  const keyword = searchParams.get('keyword') || '';
   const [open, setOpen] = useState(false);
   const [query, setQuery] = useState('');
   const debouncedQuery = useDebounce(query, 500);
-  const [isSearching, setIsSearching] = useState(false);
-
-  const { data: courses, isLoading } = useGetCourses(
-    {
-      isPublished: true,
-      keyword,
-      limit: 4
-    },
-    isSearching
-  );
 
   useEffect(() => {
     if (debouncedQuery) {
@@ -42,7 +27,6 @@ const SearchCommandMenu = () => {
         key: 'keyword',
         value: debouncedQuery
       });
-      setIsSearching(true);
       router.push(newUrl, { scroll: false });
     } else {
       const newUrl = removeKeysFromQuery({
@@ -111,46 +95,17 @@ const SearchCommandMenu = () => {
           placeholder="Type a command or search..."
         />
         <CommandList>
-          <CommandEmpty
-            className={cn(isLoading ? 'hidden' : 'py-6 text-center text-sm')}
+          <Suspense
+            fallback={
+              <div className="space-y-1 overflow-hidden px-1 py-2">
+                <Skeleton className="h-12 rounded-sm" />
+                <Skeleton className="h-12 rounded-sm" />
+                <Skeleton className="h-12 rounded-sm" />
+              </div>
+            }
           >
-            No results found.
-          </CommandEmpty>
-          {isLoading ? (
-            <div className="space-y-1 overflow-hidden px-1 py-2">
-              <Skeleton className="h-12 rounded-sm" />
-              <Skeleton className="h-12 rounded-sm" />
-              <Skeleton className="h-12 rounded-sm" />
-            </div>
-          ) : (
-            <CommandGroup
-              key={1}
-              className="space-y-2 capitalize"
-              heading={'Courses'}
-            >
-              {courses &&
-                courses?.map((item: any, index: number) => {
-                  return (
-                    <CommandItem
-                      key={item._id}
-                      className="h-14  cursor-pointer space-x-4"
-                      value={item?.title}
-                      onSelect={() =>
-                        handleSelect(() => router.push(`/courses/${item._id}`))
-                      }
-                    >
-                      <Image
-                        src={item?.thumbnail?.url}
-                        width={56}
-                        height={56}
-                        alt="Course Image"
-                      />
-                      <span className="truncate">{item?.title}</span>
-                    </CommandItem>
-                  );
-                })}
-            </CommandGroup>
-          )}
+            <SearchCourseList handleSelect={handleSelect} />
+          </Suspense>
         </CommandList>
       </CommandDialog>
     </>
